Moved admin to the new module
