+++ src/admin/error.rs
@@ -1,0 +1,1 @@
+pub use crate::s3::error::Error;
